[Silver II] Title: A → B, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/16953. A → B/A → B.cc
@@ -3,7 +3,7 @@
 
 using namespace std;
 
-long long Find(long long number, long long target, int count)
+int Find(long long number, long long target, int count)
 {
     if (number > target)
     {
@@ -23,7 +23,7 @@ int main(void)
     int A, B;
     cin >> A >> B;
 
-    long long count = Find(A, B, 1);
+    int count = Find(A, B, 1);
 
     if (count == INT_MAX) count = -1;
